refactor: change description, change Natalia photo, add greeting,
