+++ src/helpers/aboutProjectConfig.ts
@@ -1,0 +1,22 @@
+export const AboutUsInfo = [
+  {
+    title: 'Кто мы?',
+    content:
+      'Мы три космонавта-программиста: Денис — капитан нашего звездного судна, Наталья и Лера — супер помощницы, готовые к любым задачам в мире кода. Хотите узнать о нас чуть больше? Просто кликните на одного из космонавтов и ознакомьтесь с нашими\n' +
+      '          историям!'
+  },
+  {
+    title: 'О чем проект?',
+    content:
+      'Это не просто сайт, а космическое приключение, где ты можешь приобрести разные товары с Марса, Земли или Венеры. Хочешь необычный сувенир с Красной планеты? Или, может быть, кого-то пушистого с Венеры? А если ты настоящий землянин, у нас тоже найдётся что-то особенное!\n' +
+      '\n' +
+      'Каждый товар имеет необычное описание, которое заставит тебя улыбнуться, и милую картинку, которая растопит твоё сердце. Добро пожаловать в наш интернет-магазин, где космические товары и межпланетный шопинг становятся реальностью!\n' +
+      '\n' +
+      'Наш сайт создан с любовью и на основе современных технологий: React, Redux Toolkit, TypeScript и нескольких вспомогательных библиотек.'
+  },
+  {
+    title: 'Как строилась разработка?',
+    content:
+      'В нашем проекте каждый привнес свою искру таланта! Мы как команда всегда на одной волне. Лера с Натальей создали волшебные картинки и запоминающиеся описания, а Денис умело вдохновлял наши идеи анимациями и крутыми дизайнерскими штрихами. А что касается логики, то мы разделили её поровну между собой и она стала неразрывной нитью, соединяющей наши умы в гармоничный поток идей!'
+  }
+];
