+++ src/helpers/membersConfig.tsx
@@ -1,5 +1,5 @@
 import NataliaPhoto from '../assets/img/membersPhotos/Natalia.png';
-import NataliaReal from '../assets/img/membersPhotos/Natalia_real.jpg';
+import NataliaReal from '../assets/img/membersPhotos/Natalia_real.webp';
 import DenisPhoto from '../assets/img/membersPhotos/Denis.jpg';
 import DenisReal from '../assets/img/membersPhotos/Denis_real.jpg';
 import LeraPhoto from '../assets/img/membersPhotos/Lera.png';
@@ -13,6 +13,7 @@ export interface MemberData {
 }
 const NataliaDescription = (
   <>
+    <p>Привет Всем! Меня зовут Наталья.</p>
     <p>
       С ранних лет я была увлечена спортом и планировала связать свою жизнь с этой сферой. Окончила институт
       физической культуры и спорта Сибирского федерального университета в Красноярске и более 10 лет
@@ -108,28 +109,5 @@ export const members: MemberData[] = [
     description: NataliaDescription,
     photos: [NataliaPhoto, NataliaReal],
     gitHub: 'https://github.com/SplitCode'
-  }
-];
-
-export const AboutUsInfo = [
-  {
-    title: 'Кто мы?',
-    content:
-      'Мы три космонавта-программиста: Денис — капитан нашего звездного судна, Наталья и Лера — супер помощницы, готовые к любым задачам в мире кода. Хотите узнать о нас чуть больше? Просто кликните на одного из космонавтов и ознакомьтесь с нашими\n' +
-      '          историям!'
-  },
-  {
-    title: 'О чем проект?',
-    content:
-      'Это не просто сайт, а космическое приключение, где ты можешь приобрести разные товары с Марса, Земли или Венеры. Хочешь необычный сувенир с Красной планеты? Или, может быть, кого-то пушистого с Венеры? А если ты настоящий землянин, у нас тоже найдётся что-то особенное!\n' +
-      '\n' +
-      'Каждый товар имеет необычное описание, которое заставит тебя улыбнуться, и милую картинку, которая растопит твоё сердце. Добро пожаловать в наш интернет-магазин, где космические товары и межпланетный шопинг становятся реальностью!\n' +
-      '\n' +
-      'Наш сайт создан с любовью и на основе современных технологий: React, Redux Toolkit, TypeScript и нескольких вспомогательных библиотек.'
-  },
-  {
-    title: 'Как строилась разработка?',
-    content:
-      'В нашем проекте каждый привнес свою искру таланта! Мы как команда всегда на одной волне. Лера с Натальей создали волшебные картинки и запоминающиеся описания, а Денис умело вдохновлял наши идеи анимациями и крутыми дизайнерскими штрихами. А что касается логики, то мы разделили её поровну между собой, чтобы наш проект был ещё более крутым!'
   }
 ];
